Optimized stack data usage to actually support 64 bit

+++ create_task.c
@@ -12,7 +12,7 @@ int create_task(char task[]) {
 	FILE *fp;
 	fp = open_todo_file("a");
 
-	for (int count = 0;; count++) {
+	for (unsigned long count = 0;; count++) {
 		if (task[count] == '\0') break;
 		if (task[count] == '\n') {
 			fprintf(stderr, "Error: Task cannot contain line breaks.\n");
